+++ crypt-core/src/main/java/de/alpharogroup/crypto/xml/XmlEnDecryptionExtensions.java
@@ -26,24 +26,22 @@ package de.alpharogroup.crypto.xml;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.charset.Charset;
 import java.util.Map;
 
+import de.alpharogroup.crypto.file.xml.XmlDecryptionExtensions;
+import de.alpharogroup.crypto.file.xml.XmlEncryptionExtensions;
 import org.apache.commons.codec.DecoderException;
 
 import com.thoughtworks.xstream.XStream;
 
-import de.alpharogroup.crypto.hex.HexExtensions;
-import de.alpharogroup.file.read.ReadFileExtensions;
-import de.alpharogroup.file.write.WriteFileExtensions;
-import de.alpharogroup.xml.ObjectToXmlExtensions;
-import de.alpharogroup.xml.XmlToObjectExtensions;
 import lombok.NonNull;
 import lombok.experimental.UtilityClass;
 
 /**
  * The class {@link XmlEnDecryptionExtensions} provides methods for write and encrypt and read and
  * decrypt xml data.
+ * @deprecated use instead <code>XmlEncryptionExtensions</code> or <code>XmlDecryptionExtensions</code>
+ * <br>Note: will be removed in next minor release
  */
 @UtilityClass
 public class XmlEnDecryptionExtensions
@@ -65,15 +63,14 @@ public class XmlEnDecryptionExtensions
 	 *            the file to write
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
+	 * @deprecated use instead <code>XmlEncryptionExtensions</code> same name method
+	 * <br>Note: will be removed in next minor release
 	 */
 	public static <T> void writeToFileAsXmlAndHex(final @NonNull XStream xstream,
 		final @NonNull Map<String, Class<?>> aliases, final @NonNull T data, @NonNull File file)
 		throws IOException
 	{
-		String xmlString = ObjectToXmlExtensions.toXmlWithXStream(xstream, data, aliases);
-		final String hexXmlString = HexExtensions.encodeHex(xmlString, Charset.forName("UTF-8"),
-			true);
-		WriteFileExtensions.writeStringToFile(file, hexXmlString, "UTF-8");
+		XmlEncryptionExtensions.writeToFileAsXmlAndHex(xstream, aliases, data, file);
 	}
 
 	/**
@@ -93,15 +90,14 @@ public class XmlEnDecryptionExtensions
 	 *             Signals that an I/O exception has occurred.
 	 * @throws DecoderException
 	 *             is thrown if an odd number or illegal of characters is supplied
+	 * @deprecated use instead <code>XmlDecryptionExtensions</code> same name method
+	 * <br>Note: will be removed in next minor release
 	 */
 	public static <T> T readFromFileAsXmlAndHex(final @NonNull XStream xstream,
 		final @NonNull Map<String, Class<?>> aliases, final @NonNull File selectedFile)
 		throws IOException, DecoderException
 	{
-		final String hexXmlString = ReadFileExtensions.readFromFile(selectedFile);
-		String xmlString = HexExtensions.decodeHex(hexXmlString);
-
-		return XmlToObjectExtensions.toObjectWithXStream(xstream, xmlString, aliases);
+		return XmlDecryptionExtensions.readFromFileAsXmlAndHex(xstream, aliases, selectedFile);
 	}
 
 }