+++ crypt-core/src/test/java/de/alpharogroup/crypto/xml/XmlDecryptionExtensionsTest.java
@@ -28,14 +28,12 @@ import static org.testng.Assert.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
 import java.util.Map;
 
+import de.alpharogroup.crypto.file.xml.XmlDecryptionExtensions;
 import org.apache.commons.codec.DecoderException;
-import org.meanbean.test.BeanTestException;
-import org.meanbean.test.BeanTester;
+import org.junit.Test;
 import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
 
 import com.google.common.collect.BiMap;
 import com.thoughtworks.xstream.XStream;
@@ -47,9 +45,9 @@ import de.alpharogroup.crypto.obfuscation.rule.ObfuscationOperationRule;
 import de.alpharogroup.file.search.PathFinder;
 
 /**
- * The unit test class for the class {@link XmlEnDecryptionExtensions}
+ * The unit test class for the class {@link XmlDecryptionExtensions}
  */
-public class XmlEnDecryptionExtensionsTest
+public class XmlDecryptionExtensionsTest
 {
 
 	/** The aliases for the {@link XStream} object */
@@ -80,50 +78,19 @@ public class XmlEnDecryptionExtensionsTest
 	}
 
 	/**
-	 * Test method for {@link XmlEnDecryptionExtensions#readFromFileAsXmlAndHex(XStream, Map, File)}
+	 * Test method for {@link XmlDecryptionExtensions#readFromFileAsXmlAndHex(XStream, Map, File)}
 	 *
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 * @throws DecoderException
 	 *             is thrown if an odd number or illegal of characters is supplied
 	 */
-	@SuppressWarnings("deprecation")
 	@Test
 	public void testReadFromFileAsXmlAndHex() throws IOException, DecoderException
 	{
 		expected = ObfuscationOperationTestData.getFirstBiMapObfuscationOperationRules();
-		actual = XmlEnDecryptionExtensions.readFromFileAsXmlAndHex(xStream, aliases, xmlFile);
+		actual = XmlDecryptionExtensions.readFromFileAsXmlAndHex(xStream, aliases, xmlFile);
 		assertEquals(actual, expected);
 	}
 
-	/**
-	 * Test method for {@link XmlEnDecryptionExtensions} with {@link BeanTester}
-	 */
-	@SuppressWarnings("deprecation")
-	@Test(expectedExceptions = { BeanTestException.class, InvocationTargetException.class,
-			UnsupportedOperationException.class })
-	public void testWithBeanTester()
-	{
-		BeanTester beanTester = new BeanTester();
-		beanTester.testBean(XmlEnDecryptionExtensions.class);
-	}
-
-	/**
-	 * Test method for
-	 * {@link XmlEnDecryptionExtensions#writeToFileAsXmlAndHex(XStream, Map, Object, File)}
-	 *
-	 * @throws IOException
-	 *             Signals that an I/O exception has occurred.
-	 * @throws DecoderException
-	 *             is thrown if an odd number or illegal of characters is supplied
-	 */
-	@SuppressWarnings("deprecation")
-	@Test
-	public void testWriteToFileAsXmlAndHex() throws IOException, DecoderException
-	{
-		expected = ObfuscationOperationTestData.getFirstBiMapObfuscationOperationRules();
-		XmlEnDecryptionExtensions.writeToFileAsXmlAndHex(xStream, aliases, expected, xmlFile);
-		actual = XmlEnDecryptionExtensions.readFromFileAsXmlAndHex(xStream, aliases, xmlFile);
-		assertEquals(actual, expected);
-	}
 }